Fast Blake2B-256 hash on gpu

+++ gpublake2b/blake2b.cu
@@ -59,7 +59,7 @@ static const uint8_t sigma[12][16] = {
 // Compression function. "last" flag indicates last block.
 static void blake2b_compress(
     blake2b_ctx * ctx,
-    int last
+    uint32_t last
 ) {
     int i;
     uint64_t v[16];
@@ -135,10 +135,10 @@ static void blake2b_compress(
 //      Secret key (also <= 64 bytes) is optional (keylen = 0).
 int blake2b_init(
     blake2b_ctx * ctx,
-    size_t outlen,
+    uint32_t outlen,
     const void * key,
     // (keylen=0: no key)
-    size_t keylen
+    uint32_t keylen
 ) {
     size_t k;
 
@@ -186,7 +186,7 @@ void blake2b_update(
     // data
     const void * in,
     // data byte size
-    size_t inlen
+    uint32_t inlen
 ) {
     size_t k;
